+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [Feature] Add `%sqlcmd connect` to see existing connections and create new ones (#632)
 * [Fix] Improve error when passing a non-identifier to start a connection (#764)
 * [Fix] Display a warning (instead of raising an error) if the `default` connection in the `.ini` file cannot start
+* [Fix] Display a message instead of an error when `toml` isn't installed and `pyproject.toml` is found (#825)
 
 ## 0.10.0 (2023-08-19)